requirements.txt
--extra-index-url https://abetlen.github.io/llama-cpp-python-cu12.6/
llama-cpp-python --no-binary :all:
huggingface_hub